\name{GetDistance}
\alias{GetDistance}
\title{Get pdf for distance between two random points}
\usage{
  GetDistance(t, mode = 0, para)
}
\arguments{
  \item{t}{vector of points to calculate pdf for.}

  \item{mode}{\itemize{ \item 0 = square \item 1 = cube }}
}
\value{
  data
}
\description{
  to do sum of them
}
\note{
  August 17 2012
}
\examples{
GetDistance(t=1:100,mode=0,para=50)
}
\author{
  Matt Roughan, Jono Tuke
}

